// src/app/core/models/models.ts
export interface tax {
  id: number;
  date: string | number;
  economic_zone: string;
  tax: string;
  tax_rate: number;
  net: number;
  gross: number;
  amount: number;
}
// ...
export interface editcreditnotes {
  sNo: number;
  id?: number;
  name?: string;
  editcreditnotes: number;
  product: number;
  quantity: number;
  unit: string;
  rate: number;
  discount: number;
  tax: number;
  amount: number;
}